refactor: move confirm module to operations folder
