+++ src/renderer/features/operations/OperationsConfirm/Unlock/index.ts
@@ -1,0 +1,3 @@
+export { UnlockConfirmation } from './ui/UnlockConfirmation';
+export { unlockConfirmModel } from './model/unlockConfirm';
+export { type UnlockConfirm } from './model/unlockConfirm';
