+++ src/renderer/widgets/UnlockModal/index.ts
@@ -1,5 +1,2 @@
 export { UnlockModal } from './ui/UnlockModal';
-export { UnlockConfirmation } from './ui/UnlockConfirmation';
 export { unlockAggregate } from './model/unlock';
-export { unlockConfirmModel } from './model/unlockConfirm';
-export { type UnlockConfirm } from './model/unlockConfirm';
