+++ src/renderer/widgets/UnlockModal/model/unlock.ts
@@ -14,8 +14,8 @@ import { navigationModel } from '@/features/navigation';
 import { signModel } from '@/features/operations/OperationSign/model/sign-model';
 import { submitModel } from '@/features/operations/OperationSubmit';
 import { submitUtils } from '@/features/operations/OperationSubmit/lib/submit-utils';
+import { type UnlockConfirm, unlockConfirmModel } from '@/features/operations/OperationsConfirm';
 
-import { type UnlockConfirm, unlockConfirmModel } from './unlockConfirm';
 import { unlockFormAggregate } from './unlockForm';
 
 const flowStarted = createEvent();
